Lazy fonts, default to /left /tt
The custom font and the symbol font only appear in the font menu if they are mentioned in the command line, or by the user, or in a document loaded in the writer.  Similarly, the custom font is only loaded under these circumstances.  Change the default font size to 2/9 inch (except for the custom font).  Start in left half-screen mode by default.  Start with Courier New in the editor by default.