Writer font menu shows current font.  Set font other in the writer with nothing selected.
If the user selects a shape that contains a text, that text's font will be highlighted in the writer's font menus.  If the user selects a font name and/or height, that font name and/or height will be highlighted.  If the user changes pages, the contents of the page will be used to highlight a font in the font menu.  The highlighted font (in the menu) is used for newly created text.  Fixed /tt defaulting and parsing of font names from the command line.  Fixed lazy loading of custom font names in the writer.  Make the font name other and font height other menu items work in the writer, even if nothing is selected.