Add radians, mixios, and fuzzy tests.  Improved mixeds and ratios.
A mixio is like a mixed.  Internally a mixio has a whole number, some nips (which are divided by 43,200), and some trifles (which are divided by 43,200*43,200). A mixio can be converted to a string, using the same format that a mixed can be converted to a string.  Currently, mixios do not have full math functionality.  Mixios are designed to be very precise, have lots of common factors, and be less likely to have overflow errors than ratios and mixeds.  Improved mixed's string format.  Eliminated some trivial overflow conditions involving mixeds and ratios.