Support turtle's demo.
In noodle, told Windows to "DisableProcessWindowsGhosting", so that Windows will not decide that a Plain English app is not responding.  In noodle, renamed the default console to the console.  Started using alternate headers.  Included some turtle and noodle code from CAL-4700, to support its original programmer's demonstration of the turtle graphics.  Did not include the demo code.  Changed the default line color from black to the turtle's context's current color.  Moved definitions of test result, expectation, and interpretation from the status to the noodle.  In trig, renamed "precise angle" to "trig angle".  Moved definition of a fuzzy test from trig to the noodle.  Split the turtle routines into "keeper" (whose routines use the context, but leave it in its original state), "modder" (whose routines modify the context), and "turtle" (whose routines do not use the context).