Added batch script to measure lengths of source code files.
Outputs a text file with one row for each source code file, listing the number of lines of source code, just comments, blank lines, total lines, and an estimate of the number of lines that would need to be added to wrap all lines to fit in an 80-column work area.  Also outputs a row of totals.  Does not output rows for "reamish" files (the writer's save format), nor for lexicons, nor for files with extensions.  Due to batch script limitations, the noodle's Osmosian font resource is estimated to need 102 line wraps, instead of about 900 line wraps.