Improved unicode support.  Used unicode in the desktop's tabstrip and trio.
Displayed unicode in the desktop's trio, and as abbreviation characters in tabstrips with proportional fonts.  Corrected conversion from ASCII to Unicode UTF-16LE.  In the noodle, stored the "U+" prefix in the "code point" data structure.  Tested string-widths in a way that handles changes that Microsoft deployed in Windows 10 Build 1703.