Treat clear color specially.  In turtle, fade a line.  Minor refactoring of colors.
Changed how "to decide if a color is clear" to just treat the as initialized clear color as clear, with hue=-1, saturation=0, lightness=0, instead of any color with a negative hue.  For purposes of adding colors, the clear color does not modify the other color.
In the turtle, added "to fade a line a length long from a color to another color".
Moved two ratio deciders from colors to the noodle.  When converting a color to an rgb, slightly expanded the range of colors used in the intermediate calculations.  De-duplicated code for finding a color between two colors.